refactor model for test/prod health
Still trying to find the real shape of the types.

I think this is the easiest for rendering and still
maintains its descriptive quality on the type level.

A Status is made of a service, of type Service and it's
testHealth and prodHealth, both of type Condition.

Environment is now redundant.